smyrna: make 'viewport_camera.pan{x|y}' 'double's
This is part of a series of changes to standardize on `double` arithmetic in
Smyrna. This removes visual noise and improves accuracy. The computational cost
of `double` vs `float` arithmetic is no longer a significant factor in
contemporary environments.

+++ cmd/smyrna/smyrnadefs.h
@@ -247,8 +247,8 @@ enum {SCHEMACOUNT = 4}; // number of colors
     typedef struct _ViewInfo {
 	systemgraphs systemGraphs;
 	/*view variables */
-	float panx;
-	float pany;
+	double panx;
+	double pany;
 	float zoom;
 
 	/*clipping coordinates, to avoid unnecesarry rendering */
